requirements.txt
streamlit==1.29.0
google-generativeai>=0.8.0
google-cloud-bigquery==3.14.0
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
Pillow==10.1.0
python-dotenv==1.0.0
pandas==2.1.3
reportlab==4.0.9
python-dateutil==2.8.2
google-auth==2.25.2
google-api-core==2.15.0
